1차배포

+++ src/like/like.service.ts
@@ -1,10 +1,19 @@
-import { BadRequestException, Injectable } from '@nestjs/common';
+import {
+  BadRequestException,
+  ConflictException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { CommentLike } from './entities/comment-like.entity';
 import { Repository } from 'typeorm';
 import { PostLike } from './entities/post-like.entity';
 import { Post } from 'src/post/entities/post.entity';
 import { Comment } from 'src/comment/entities/comment.entity';
+import { COMMENT_MESSAGE } from 'src/constants/comment-message.constant';
+import { POST_MESSAGE } from 'src/constants/post-message.constant';
+import { PointService } from 'src/point/point.service';
+import { PointType } from 'src/point/types/point.type';
 
 @Injectable()
 export class LikeService {
@@ -16,39 +25,63 @@ export class LikeService {
     @InjectRepository(PostLike)
     private readonly postLikeRepository: Repository<PostLike>,
     @InjectRepository(CommentLike)
-    private readonly commentLikeRepository: Repository<CommentLike>
+    private readonly commentLikeRepository: Repository<CommentLike>,
+    private readonly pointService: PointService
   ) {}
 
+  // 댓글 좋아요 조회 API
   async getCommentLikes(commentId: number) {
     const existingComment = await this.commentRepository.findOneBy({
       id: commentId,
     });
+
+    // 댓글 존재 확인
     if (!existingComment) {
-      throw new BadRequestException('댓글이 존재하지 않습니다.');
+      throw new NotFoundException(COMMENT_MESSAGE.COMMENT.NOT_FOUND);
     }
 
+    // 댓글 좋아요 카운트
     const count = await this.commentLikeRepository.countBy({ commentId });
 
     return count;
   }
 
+  // 댓글 좋아요 생성 API
   async createCommentLike(userId, commentId) {
     const existingComment = await this.commentRepository.findOneBy({
       id: commentId,
     });
+    const commentLike = await this.commentLikeRepository.findOneBy({
+      userId: userId,
+      commentId: commentId,
+    });
+
+    // 댓글 존재 확인
     if (!existingComment) {
-      throw new BadRequestException('댓글이 존재하지 않습니다.');
-    } else if ((existingComment.userId = userId)) {
-      throw new BadRequestException('자신의 댓글에 좋아요를 누를 수 없습니다.');
+      throw new NotFoundException(COMMENT_MESSAGE.COMMENT.NOT_FOUND);
     }
 
-    const commentLike = await this.commentLikeRepository.findOneBy({
-      userId,
-      commentId,
-    });
-    if (commentLike) {
-      throw new BadRequestException('이미 좋아요를 누른 댓글입니다.');
+    // 댓글 작성자 ID 비교 후 동일 유저이면 좋아요 금지
+    else if (existingComment.userId === userId) {
+      throw new BadRequestException(
+        COMMENT_MESSAGE.LIKE.CREATE.FAILURE.NO_SELF
+      );
     }
+
+    // 댓글 좋아요 눌렀었는지 확인
+    if (commentLike) {
+      throw new BadRequestException(
+        COMMENT_MESSAGE.LIKE.CREATE.FAILURE.ALREADY
+      );
+    }
+
+    // 댓글 좋아요 생성 포인트 지급
+    const isValidPoint = await this.pointService.validatePointLog(
+      userId,
+      PointType.COMMENT_LIKE
+    );
+    if (isValidPoint)
+      this.pointService.savePointLog(userId, PointType.COMMENT_LIKE, true);
 
     await this.commentLikeRepository.save({
       userId,
@@ -56,13 +89,28 @@ export class LikeService {
     });
   }
 
+  // 댓글 좋아요 삭제 API
   async deleteCommentLike(userId, commentId) {
     const existingComment = await this.commentRepository.findOneBy({
       id: commentId,
     });
+    const commentLike = await this.commentLikeRepository.findOneBy({
+      userId: userId,
+      commentId: commentId,
+    });
+
+    // 댓글 존재 확인
     if (!existingComment) {
-      throw new BadRequestException('댓글이 존재하지 않습니다.');
+      throw new NotFoundException(COMMENT_MESSAGE.COMMENT.NOT_FOUND);
     }
+
+    // 내 좋아요 남아있는지 확인
+    if (!commentLike) {
+      throw new NotFoundException(COMMENT_MESSAGE.LIKE.DELETE.FAILURE.NO_LIKE);
+    }
+
+    // 댓글 삭제로 포인트 차감
+    this.pointService.savePointLog(userId, PointType.COMMENT_LIKE, false);
 
     await this.commentLikeRepository.delete({
       userId,
@@ -70,12 +118,15 @@ export class LikeService {
     });
   }
 
+  // 게시글 좋아요 조회 API
   async getPostLikes(postId) {
     const existingPost = await this.postRepository.findOneBy({
       id: postId,
     });
+
+    // 게시글 존재 확인
     if (!existingPost) {
-      throw new BadRequestException('게시글이 존재하지 않습니다.');
+      throw new NotFoundException(POST_MESSAGE.POST.NOT_FOUND);
     }
 
     const count = await this.postLikeRepository.countBy({ postId });
@@ -83,25 +134,39 @@ export class LikeService {
     return count;
   }
 
+  // 게시글 좋아요 생성 API
   async createPostLike(userId, postId) {
     const existingPost = await this.postRepository.findOneBy({
       id: postId,
     });
-    if (!existingPost) {
-      throw new BadRequestException('게시글이 존재하지 않습니다.');
-    } else if ((existingPost.userId = userId)) {
-      throw new BadRequestException(
-        '자신의 게시글에 좋아요를 누를 수 없습니다.'
-      );
-    }
 
     const postLike = await this.postLikeRepository.findOneBy({
       userId,
       postId,
     });
-    if (postLike) {
-      throw new BadRequestException('이미 좋아요를 누른 댓글입니다.');
+
+    // 게시글 존재 확인
+    if (!existingPost) {
+      throw new NotFoundException(POST_MESSAGE.POST.NOT_FOUND);
     }
+
+    // 게시글 작성자 ID 비교 후 동일 유저이면 좋아요 금지
+    else if (existingPost.userId == userId) {
+      throw new BadRequestException(POST_MESSAGE.LIKE.CREATE.FAILURE.NO_SELF);
+    }
+
+    // 게시글 좋아요 눌렀었는지 확인
+    if (postLike) {
+      throw new BadRequestException(POST_MESSAGE.LIKE.CREATE.FAILURE.ALREADY);
+    }
+
+    // 게시글 좋아요 생성 포인트 지급
+    const isValidPoint = await this.pointService.validatePointLog(
+      userId,
+      PointType.POST_LIKE
+    );
+    if (isValidPoint)
+      this.pointService.savePointLog(userId, PointType.POST_LIKE, true);
 
     await this.postLikeRepository.save({
       userId,
@@ -109,13 +174,28 @@ export class LikeService {
     });
   }
 
+  // 게시글 좋아요 삭제 API
   async deletePostLike(userId, postId) {
     const existingPost = await this.postRepository.findOneBy({
       id: postId,
     });
+    const postLike = await this.postLikeRepository.findOneBy({
+      userId,
+      postId,
+    });
+
+    // 게시글 존재 확인
     if (!existingPost) {
-      throw new BadRequestException('댓글이 존재하지 않습니다.');
+      throw new NotFoundException(POST_MESSAGE.POST.NOT_FOUND);
     }
+
+    // 내 좋아요 남아있는지 확인
+    if (!postLike) {
+      throw new NotFoundException(POST_MESSAGE.LIKE.DELETE.FAILURE.NO_LIKE);
+    }
+
+    // 게시글 삭제로 포인트 차감
+    this.pointService.savePointLog(userId, PointType.POST_LIKE, false);
 
     await this.postLikeRepository.delete({
       userId,
